requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud Platform
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-functions==1.16.0
google-cloud-aiplatform==1.38.1
google-cloud-monitoring==2.16.0
google-cloud-logging==3.8.0

# Data Processing - Updated apache-beam for better PEP 517 support
apache-beam[gcp]==2.52.0
pandas==2.1.3
numpy>=1.14.3,<1.25.0
pyarrow>=3.0.0,<12.0.0

# Image Processing and ML
opencv-python==4.8.1.78
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0
scikit-image==0.22.0

# Database and Caching
redis==5.0.1
sqlalchemy==2.0.23

# HTTP and API
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.38.0

# Configuration and Environment
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Testing (for development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Data Validation
marshmallow==3.20.1
cerberus==1.3.5

# Async and Concurrency
aiofiles==23.2.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
# Note: uuid package removed - Python 3.7+ has built-in uuid module
# Note: asyncio removed - it's part of Python standard library

# Data Integration APIs
sodapy==1.5.2
cenpy==1.0.1
fredapi==0.5.1
googlemaps==4.10.0
